changed constant from double to float

+++ src/main/java/core/basesyntax/MyHashMap.java
@@ -9,7 +9,7 @@ import java.util.Objects;
  */
 public class MyHashMap<K, V> implements MyMap<K, V> {
     private static final int DEFAULT_CAPACITY = 16;
-    private static final double LOAD_FACTOR = 0.75;
+    private static final float LOAD_FACTOR = 0.75f;
     private int size = 0;
     private int currentCapacity;
     private Node<K, V>[] hashMapList;
